Fix document download identifier type

+++ src/java/controller/ControllerServlet.java
@@ -596,7 +596,7 @@ public class ControllerServlet extends HttpServlet {
     protected void handleDocument(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
         String[] split = request.getPathInfo().split("[/-]");
-        Integer documentId = Integer.parseInt(split[1]);
+        Long documentId = Long.parseLong(split[1]);
        
         FacesContext fc = FacesContext.getCurrentInstance();
         Document document = documentFacade.find(documentId);
